+++ concourse-ete-test-core/src/main/java/com/cinchapi/concourse/server/ManagedConcourseServer.java
@@ -1213,23 +1213,6 @@ public class ManagedConcourseServer {
         }
 
         @Override
-        public Set<Long> find(Criteria criteria, Page page) {
-            return invoke("find", Criteria.class, Page.class).with(criteria,
-                    page);
-        }
-
-        @Override
-        public Set<Long> find(Object criteria) {
-            return invoke("find", Object.class).with(criteria);
-        }
-
-        @Override
-        public Set<Long> find(Object criteria, Page page) {
-            return invoke("find", Object.class, Page.class).with(criteria,
-                    page);
-        }
-
-        @Override
         public Set<Long> find(String ccl) {
             return invoke("find", String.class).with(ccl);
         }
@@ -1474,35 +1457,6 @@ public class ManagedConcourseServer {
 
         @Override
         public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-                Object criteria) {
-            return invoke("get", Collection.class, Object.class).with(keys,
-                    criteria);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-                Object criteria, Page page) {
-            return invoke("get", Collection.class, Object.class, Page.class)
-                    .with(keys, criteria, page);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-                Object criteria, Timestamp timestamp) {
-            return invoke("get", Collection.class, Object.class,
-                    Timestamp.class).with(keys, criteria, timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-                Object criteria, Timestamp timestamp, Page page) {
-            return invoke("get", Collection.class, Object.class,
-                    Timestamp.class, Page.class).with(keys, criteria, timestamp,
-                            page);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
                 String ccl) {
             return invoke("get", Collection.class, String.class).with(keys,
                     ccl);
@@ -1546,37 +1500,6 @@ public class ManagedConcourseServer {
                 Timestamp timestamp) {
             return invoke("get", Criteria.class, Timestamp.class).with(criteria,
                     timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Criteria criteria,
-                Timestamp timestamp, Page page) {
-            return invoke("get", Criteria.class, Timestamp.class, Page.class)
-                    .with(criteria, timestamp, page);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Object criteria) {
-            return invoke("get", Object.class).with(criteria);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Object criteria, Page page) {
-            return invoke("get", Object.class, Page.class).with(criteria, page);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Object criteria,
-                Timestamp timestamp) {
-            return invoke("get", Object.class, Timestamp.class).with(criteria,
-                    timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, T>> get(Object criteria,
-                Timestamp timestamp, Page page) {
-            return invoke("get", Object.class, Timestamp.class, Page.class)
-                    .with(criteria, timestamp, page);
         }
 
         @Override
@@ -1652,32 +1575,6 @@ public class ManagedConcourseServer {
         public <T> T get(String key, long record, Timestamp timestamp) {
             return invoke("get", String.class, long.class, Timestamp.class)
                     .with(key, record, timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, T> get(String key, Object criteria) {
-            return invoke("get", String.class, Object.class).with(key,
-                    criteria);
-        }
-
-        @Override
-        public <T> Map<Long, T> get(String key, Object criteria, Page page) {
-            return invoke("get", String.class, Object.class, Page.class)
-                    .with(key, criteria, page);
-        }
-
-        @Override
-        public <T> Map<Long, T> get(String key, Object criteria,
-                Timestamp timestamp) {
-            return invoke("get", String.class, Object.class, Timestamp.class)
-                    .with(key, criteria, timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, T> get(String key, Object criteria,
-                Timestamp timestamp, Page page) {
-            return invoke("get", String.class, Object.class, Timestamp.class,
-                    Page.class).with(key, criteria, timestamp, page);
         }
 
         @Override
@@ -2127,12 +2024,6 @@ public class ManagedConcourseServer {
         }
 
         @Override
-        public Set<Long> search(String key, String query, Page page) {
-            return invoke("search", String.class, String.class, Page.class)
-                    .with(key, query, page);
-        }
-
-        @Override
         public Map<Long, Map<String, Set<Object>>> select(
                 Collection<Long> records) {
             return invoke("select", Collection.class).with(records);
@@ -2236,36 +2127,6 @@ public class ManagedConcourseServer {
 
         @Override
         public <T> Map<Long, Map<String, Set<T>>> select(
-                Collection<String> keys, Object criteria) {
-            return invoke("select", Collection.class, Object.class).with(keys,
-                    criteria);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                Collection<String> keys, Object criteria, Page page) {
-            return invoke("select", Collection.class, Object.class, Page.class)
-                    .with(keys, criteria, page);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                Collection<String> keys, Object criteria, Timestamp timestamp) {
-            return invoke("select", Collection.class, Object.class,
-                    Timestamp.class).with(keys, criteria, timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
-                Collection<String> keys, Object criteria, Timestamp timestamp,
-                Page page) {
-            return invoke("select", Collection.class, Object.class,
-                    Timestamp.class, Page.class).with(keys, criteria, timestamp,
-                            page);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(
                 Collection<String> keys, String ccl) {
             return invoke("select", Collection.class, String.class).with(keys,
                     ccl);
@@ -2330,32 +2191,6 @@ public class ManagedConcourseServer {
                 Timestamp timestamp) {
             return invoke("select", long.class, Timestamp.class).with(record,
                     timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(Object criteria) {
-            return invoke("select", Object.class).with(criteria);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(Object criteria,
-                Page page) {
-            return invoke("select", Object.class, Page.class).with(criteria,
-                    page);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(Object criteria,
-                Timestamp timestamp) {
-            return invoke("select", Object.class, Timestamp.class)
-                    .with(criteria, timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Map<String, Set<T>>> select(Object criteria,
-                Timestamp timestamp, Page page) {
-            return invoke("select", Object.class, Timestamp.class, Page.class)
-                    .with(criteria, timestamp, page);
         }
 
         @Override
@@ -2435,33 +2270,6 @@ public class ManagedConcourseServer {
         public <T> Set<T> select(String key, long record, Timestamp timestamp) {
             return invoke("select", String.class, long.class, Timestamp.class)
                     .with(key, record, timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(String key, Object criteria) {
-            return invoke("select", String.class, Object.class).with(key,
-                    criteria);
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(String key, Object criteria,
-                Page page) {
-            return invoke("select", String.class, Object.class, Page.class)
-                    .with(key, criteria, page);
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(String key, Object criteria,
-                Timestamp timestamp) {
-            return invoke("select", String.class, Object.class, Timestamp.class)
-                    .with(key, criteria, timestamp);
-        }
-
-        @Override
-        public <T> Map<Long, Set<T>> select(String key, Object criteria,
-                Timestamp timestamp, Page page) {
-            return invoke("select", String.class, Object.class, Timestamp.class,
-                    Page.class).with(key, criteria, timestamp, page);
         }
 
         @Override