+++ concourse-import/src/main/java/com/cinchapi/concourse/importer/debug/ImportDryRunConcourse.java
@@ -306,21 +306,6 @@ public class ImportDryRunConcourse extends Concourse {
     }
 
     @Override
-    public Set<Long> find(Criteria criteria, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Set<Long> find(Object criteria) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Set<Long> find(Object criteria, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
     public Set<Long> find(String ccl) {
         throw new UnsupportedOperationException();
     }
@@ -524,30 +509,6 @@ public class ImportDryRunConcourse extends Concourse {
 
     @Override
     public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-            Object criteria) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-            Object criteria, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-            Object criteria, Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
-            Object criteria, Timestamp timestamp, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Collection<String> keys,
             String ccl) {
         throw new UnsupportedOperationException();
     }
@@ -583,34 +544,6 @@ public class ImportDryRunConcourse extends Concourse {
     @Override
     public <T> Map<Long, Map<String, T>> get(Criteria criteria,
             Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Criteria criteria,
-            Timestamp timestamp, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Object criteria) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Object criteria, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Object criteria,
-            Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, T>> get(Object criteria,
-            Timestamp timestamp, Page page) {
         throw new UnsupportedOperationException();
     }
 
@@ -676,28 +609,6 @@ public class ImportDryRunConcourse extends Concourse {
 
     @Override
     public <T> T get(String key, long record, Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, T> get(String key, Object criteria) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, T> get(String key, Object criteria, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, T> get(String key, Object criteria,
-            Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, T> get(String key, Object criteria,
-            Timestamp timestamp, Page page) {
         throw new UnsupportedOperationException();
     }
 
@@ -1074,11 +985,6 @@ public class ImportDryRunConcourse extends Concourse {
     }
 
     @Override
-    public Set<Long> search(String key, String query, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
     public Map<Long, Map<String, Set<Object>>> select(
             Collection<Long> records) {
         throw new UnsupportedOperationException();
@@ -1164,30 +1070,6 @@ public class ImportDryRunConcourse extends Concourse {
 
     @Override
     public <T> Map<Long, Map<String, Set<T>>> select(Collection<String> keys,
-            Object criteria) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Collection<String> keys,
-            Object criteria, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Collection<String> keys,
-            Object criteria, Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Collection<String> keys,
-            Object criteria, Timestamp timestamp, Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Collection<String> keys,
             String ccl) {
         throw new UnsupportedOperationException();
     }
@@ -1240,29 +1122,6 @@ public class ImportDryRunConcourse extends Concourse {
 
     @Override
     public Map<String, Set<Object>> select(long record, Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Object criteria) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Object criteria,
-            Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Object criteria,
-            Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Map<String, Set<T>>> select(Object criteria,
-            Timestamp timestamp, Page page) {
         throw new UnsupportedOperationException();
     }
 
@@ -1329,23 +1188,6 @@ public class ImportDryRunConcourse extends Concourse {
 
     @Override
     public <T> Set<T> select(String key, long record, Timestamp timestamp) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Set<T>> select(String key, Object criteria) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Set<T>> select(String key, Object criteria,
-            Page page) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> Map<Long, Set<T>> select(String key, Object criteria,
-            Timestamp timestamp) {
         throw new UnsupportedOperationException();
     }